:recycle: [refactor] 예외 처리 조건문 메소드 분리
