+++ docs/README.md
@@ -1,5 +1,5 @@
 ## 🚧 수정 사항과 공부할 내용 리스트 정리
-- [ ] 복잡한 if 문의 조건식과 반복되는 if문 메소드로 분리 + 주석 제거   
+- [x] 복잡한 if 문의 조건식과 반복되는 if문 메소드로 분리 + 주석 제거   
 --> 조건문 분해(Decompose Conditional),  https://velog.io/@hi_potato/Refactoring-Simplifying-Conditional-Expressions
 ~~--> 함수형 인터페이스, 동작 매개 변수화 -> 람다식 사용~~
 - [ ] 매개변수에 의한 예외 말고 또다른 예외 찾기
